first commit
[Practice class/object]

Design and implement a class called Box that contains instance data
that represent the height,width,and depth of the box.Also include a
boolean variable called full as instance data that represent whether
the box is full or not.Define the Box constructor to accept and
initialize the height,width,and depth of the box.Each newly created Box
is empty (the constructor should initialize full to false). Include
getter and setter methods for all instance data.Include a toString
method that returns a one-line description of the box.Create a driver
class called BoxTest,whose main method instantiates and updates several
Box object.